# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CreateChannelRequest:
    properties:
      is_private:
        description: If true, creator will be added as member
        type: boolean
      name:
        type: string
      project_id:
        type: string
    required:
    - name
    - project_id
    type: object
...
  models.Channel:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      is_private:
        description: Computed field, not stored in DB
        type: boolean
      members:
        items:
          $ref: '#/definitions/models.User'
        type: array
      name:
        type: string
      project:
        $ref: '#/definitions/models.Project'
      project_id:
        type: string
      updated_at:
        type: string
    type: object
  models.ChannelMember:
    properties:
      channel:
        $ref: '#/definitions/models.Channel'
      channel_id:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
...
  models.Project:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.RoleType:
    enum:
    - admin
    - manager
    - user
    - external
    type: string
    x-enum-comments:
      RoleAdmin: Can manage everything
      RoleExternal: Can only participate in public channels
      RoleManager: Can manage everything except users
      RoleUser: Can participate in assigned projects/channels, create 1:1 chats
    x-enum-varnames:
    - RoleAdmin
    - RoleManager
    - RoleUser
    - RoleExternal
  models.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      system_role:
        allOf:
        - $ref: '#/definitions/models.RoleType'
        description: Default system role
      updated_at:
        type: string
    type: object
  models.UserResponse:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      system_role:
        $ref: '#/definitions/models.RoleType'
      updated_at:
        type: string
    type: object
...
host: localhost:30000
info:
  contact:
    email: support@thothix.com
    name: API Support
    url: http://www.thothix.com/support
  description: API per la piattaforma di messaggistica aziendale Thothix
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Thothix API
  version: "1.0"
paths:
  /api/v1/auth/me:
    get:
      consumes:
      - application/json
      description: Get the current authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /api/v1/auth/sync:
    post:
      consumes:
      - application/json
      description: Synchronize user data from Clerk to local database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Sync user from Clerk
      tags:
      - auth
  /api/v1/channels:
    get:
      consumes:
      - application/json
      description: Get a list of all channels accessible to the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Channel'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all channels for user
      tags:
      - channels
    post:
      consumes:
      - application/json
      description: Create a new channel for a project
      parameters:
      - description: Channel data
        in: body
        name: channel
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateChannelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Channel'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new channel
      tags:
      - channels
  /api/v1/channels/{id}:
    get:
      consumes:
      - application/json
      description: Get a single channel by its ID
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Channel'
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get channel by ID
      tags:
      - channels
  /api/v1/channels/{id}/join:
    post:
      consumes:
      - application/json
      description: Join a public channel or accept invite to private channel
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ChannelMember'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Join a channel
      tags:
      - channels
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"